requirements.txt
appnope==0.1.0
astroid==1.6.4
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.0
Django==2.0.5
django-extensions==2.0.7
djangorestframework==3.8.2
idna==2.6
isort==4.3.4
itypes==1.1.0
jedi==0.12.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
parso==0.2.0
pep8==1.7.1
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pylint==1.9.1
pytz==2018.4
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
wrapt==1.10.11
